// opencti-platform/opencti-front/src/private/components/common/stix_domain_objects/diamond/types/nodes/diamondEnums.ts
export enum DiamondNodeEnum {
  adversary = 'adversary',
  infrastructure = 'infrastructure',
  capabilities = 'capabilities',
}

export enum DiamondEntityEnum {
  threatActorGroup = 'Threat-Actor-Group',
  threatActorIndividual = 'Threat-Actor-Individual',
  intrusionSet = 'Intrusion-Set',
  campaign = 'Campaign',
  malware = 'Malware',
  channel = 'Channel',
  tool = 'Tool',
  incident = 'Incident',
  infrastructure = 'Infrastructure',
}